+++ export/requirements.txt
@@ -1,6 +1,3 @@
 google-cloud-storage
-aiohttp
 pyyaml
 duckdb
-pandas
-datetime
