+++ ingest/requirements.txt
@@ -1,0 +1,4 @@
+aiohttp
+pyyaml
+duckdb
+pandas
